Fix problems
- fix errors
- add import to UI
