+++ core/preferences.json
@@ -1,0 +1,3 @@
+{
+    "language": 0
+}
